refactor(cmd): split preview.go by responsibility
Splits the 1,191-line pkg/cmd/internal/templates/preview.go into eight
focused files within the same package. Pure file-organization refactor:
no logic, signature, or visibility changes.

New layout:
- preview.go           - SetPreviewMode, previewTime, PreviewType constants
- preview_dispatch.go  - PreviewCLIOutput switch (the big router)
- preview_samples.go   - shared sample data (samplePlanChanges,
                          samplePlanLintViolations, seqDDLs)
- preview_plan.go      - plan / multi-env plan renderers
- preview_lock.go      - lock-state renderers
- preview_status.go    - state / status-list / status-history renderers
- preview_lint.go      - lint + unsafe renderers
- preview_comment.go   - PR-comment + paired-aggregate renderers
                          (incl. printSections helper)

Benefits:
- Each file now has a single, nameable responsibility instead of one
  1.2k-LOC mixed file.
- Sample data (samplePlanChanges, samplePlanLintViolations, seqDDLs) is
  isolated in preview_samples.go, separating fixture generation from
  rendering — the original goal of the "rationalize preview / rendering
  packages so sample-data generation and rendering logic are not
  intertwined" follow-up.
- Easier to add a new preview category (one new file) or extend an
  existing one without touching the dispatcher.
- Smaller merge-conflict surface area on the preview package.
- Zero behavioural risk: nothing was renamed, re-exported, or rewired
  — only moved.

Validation:
- go build ./...
- go test -count=1 -race ./pkg/cmd/...           (all green)
- golangci-lint run --new-from-rev=origin/main
    ./pkg/cmd/...                                (0 issues)
- go run ./pkg/cmd preview all                   (1879 lines, unchanged)
